requirements.txt
pyyaml
scrapy
scrapy-playwright
pandas
numpy
pytest
beautifulsoup4
requests
openpyxl
bsuite
dm-acme
tensorflow
tensorflow-probability
dm-sonnet
tqdm